Image editor: keep reducer action type with core types
