+++ packages/media-editor/src/image-editor/core/state.ts
@@ -1,8 +1,7 @@
 /**
  * Internal dependencies
  */
-import type { CropperAction } from './actions';
-import type { CropperState, TransformOperation } from './types';
+import type { CropperAction, CropperState, TransformOperation } from './types';
 import { DEFAULT_STATE, MAX_ZOOM } from './constants';
 import { normalizeRotation, degreesToRadians } from './math/rotation';
 import { restrictPanZoom, restrictCropRect } from './containment';
